+++ snips_nlu/tests/test_config.py
@@ -2,14 +2,20 @@
 from __future__ import unicode_literals
 
 import unittest
+from copy import deepcopy
 
+from snips_nlu_ontology import get_all_languages
+
+from snips_nlu import SnipsNLUEngine
+from snips_nlu.constants import LANGUAGE, RES_INTENT, RES_INTENT_NAME
+from snips_nlu.default_configs import DEFAULT_CONFIGS
 from snips_nlu.intent_classifier import LogRegIntentClassifier
 from snips_nlu.pipeline.configs import (
     CRFSlotFillerConfig, SlotFillerDataAugmentationConfig,
     LogRegIntentClassifierConfig, IntentClassifierDataAugmentationConfig,
     FeaturizerConfig, NLUEngineConfig, ProbabilisticIntentParserConfig,
     DeterministicIntentParserConfig)
-from snips_nlu.tests.utils import SnipsTest
+from snips_nlu.tests.utils import SnipsTest, WEATHER_DATASET
 
 
 class TestConfig(SnipsTest):
@@ -18,6 +24,7 @@ class TestConfig(SnipsTest):
         config_dict = {
             "min_utterances": 3,
             "noise_factor": 2,
+            "add_builtin_entities_examples": False,
             "unknown_word_prob": 0.1,
             "unknown_words_replacement_string": "foobar",
         }
@@ -33,7 +40,8 @@ class TestConfig(SnipsTest):
         # Given
         config_dict = {
             "min_utterances": 42,
-            "capitalization_ratio": 0.66
+            "capitalization_ratio": 0.66,
+            "add_builtin_entities_examples": False
         }
 
         # When
@@ -161,6 +169,23 @@ class TestConfig(SnipsTest):
 
         # Then
         self.assertDictEqual(config_dict, serialized_config)
+
+    def test_default_configs_should_work(self):
+        # Given
+        dataset = deepcopy(WEATHER_DATASET)
+
+        for language in get_all_languages():
+            # When
+            config = DEFAULT_CONFIGS.get(language)
+            self.assertIsNotNone(config, "Missing default config for '%s'"
+                                 % language)
+            dataset[LANGUAGE] = language
+            engine = SnipsNLUEngine(config).fit(dataset)
+            result = engine.parse("Please give me the weather in Paris")
+
+            # Then
+            intent_name = result[RES_INTENT][RES_INTENT_NAME]
+            self.assertEqual("SearchWeatherForecast", intent_name)
 
 
 if __name__ == '__main__':