+++ snips_nlu/tests/test_features_utils.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 
 from snips_nlu.constants import NGRAM, TOKEN_INDEXES
-from snips_nlu.slot_filler.features_utils import get_all_ngrams, get_shape
+from snips_nlu.slot_filler.features_utils import get_all_ngrams
 from snips_nlu.tests.utils import SnipsTest
 
 
@@ -35,19 +35,3 @@ class TestFeaturesUtils(SnipsTest):
         ]
 
         self.assertListEqual(expected_ngrams, ngrams)
-
-    def test_get_shape(self):
-        # Given
-        data = [
-            ("tragédie", "xxx"),
-            ("Tragédie", "Xxx"),
-            ("TRAGÉDIE", "XXX"),
-            ("traGédie", "xX")
-        ]
-
-        for text, expected_capitalization in data:
-            # When
-            capitalization = get_shape(text)
-
-            # Then
-            self.assertEqual(capitalization, expected_capitalization)